refactoring annotations converter into modular functions; fix-precommit
