+++ src/tissue_map_tools/data_model/annotations.py
@@ -1,4 +1,4 @@
-from typing import Literal, Any, Self
+from typing import Literal, Any, Self, cast
 from pydantic import (
     BaseModel,
     Field,
@@ -14,6 +14,18 @@ from pathlib import PurePosixPath
 import re
 import struct
 from pathlib import Path
+from scipy.spatial import KDTree
+import matplotlib.pyplot as plt
+import numpy as np
+from dask.dataframe import DataFrame as DaskDataFrame
+import pandas as pd
+from numpy.random import default_rng
+import itertools
+from numpy.typing import NDArray
+
+RNG = default_rng(42)
+
+PRINT_DEBUG = False
 
 
 class AnnotationProperty(BaseModel):
@@ -553,3 +565,258 @@ def read_related_object_id_index(
                 relationship_data[object_id] = decoded_data
         all_relationships_data[rel.id] = relationship_data
     return all_relationships_data
+
+
+class GridLevel:
+    level: int
+    grid_shape: list[int]
+    chunk_size: NDArray[np.float64]
+    limit: int
+
+    def __init__(
+        self,
+        level: int,
+        grid_shape: list[int],
+        mins: NDArray[np.float64],
+        maxs: NDArray[np.float64],
+        limit: int,
+        parent_cells: list[tuple[int, int, int]],
+        parent_grid_shape: list[int],
+    ) -> None:
+        self.level = level
+        self.grid_shape = grid_shape
+        self.mins = mins
+        self.maxs = maxs
+        self.limit = limit
+
+        # derived quantities
+        self.sizes = np.array(maxs) - np.array(mins)
+        self.chunk_size = self.sizes / np.array(self.grid_shape)
+        self.cells: list[tuple[int, int, int]] = []
+
+        # quantities set later
+        self.populated_cells: dict[tuple[int, int, int], NDArray[np.float64]] = {}
+
+        for parent_cell in parent_cells:
+            new_cells_by_dim: dict[int, list[int]] = {}
+            for dim in range(3):
+                index = parent_cell[dim]
+                factor = grid_shape[dim] // parent_grid_shape[dim]
+                if factor == 1:
+                    new_cells_by_dim[dim] = [index]
+                else:
+                    new_cells_by_dim[dim] = [index * factor, index * factor + 1]
+            new_cells = cast(
+                list[tuple[int, int, int]],
+                itertools.product(*new_cells_by_dim.values()),
+            )
+            self.cells.extend(new_cells)
+
+    def iter_full_grid(self):
+        for i in range(self.grid_shape[0]):
+            for j in range(self.grid_shape[1]):
+                for k in range(self.grid_shape[2]):
+                    yield (i, j, k)
+
+    def iter_cells(self):
+        for i, j, k in self.cells:
+            yield (i, j, k)
+
+    def centroid(self, index: tuple[int, int, int]) -> NDArray[np.float64]:
+        """Calculate the centroid of the grid cell."""
+        return np.array(index) * self.chunk_size + self.chunk_size / 2 + self.mins
+
+    def get_next_grid_shape(self) -> list[int]:
+        """Get the shape of the next grid level so that we get isotropic chunks.
+
+        Notes
+        -----
+        The specs say: "each component of chunk_size of each successively level
+        should be either equal to, or half of, the corresponding component of the
+        prior level chunk_size, whichever results in a more spatially isotropic
+        chunk."
+
+        We implement this as follows: if a chunk lenght for a given axis is half
+        (or less) than the size of any other chunk, then we leave this axis as is,
+        otherwise we divide the chunk size by 2.
+        """
+        next_grid_shape = self.grid_shape.copy()
+        for i in range(3):
+            if any(
+                [
+                    self.chunk_size[i] * 2 <= self.chunk_size[j].item()
+                    for j in range(3)
+                    if j != i
+                ]
+            ):
+                continue
+            else:
+                next_grid_shape[i] *= 2
+        return next_grid_shape
+
+
+def compute_spatial_index(
+    points: DaskDataFrame | pd.DataFrame,
+    limit: int = 1000,
+    starting_grid_shape: list[int] | None = None,
+) -> dict[int, GridLevel]:
+    # TODO: only points are supported at the moment, not lines, axis-aligned bounding
+    #  boxes and ellipsoids
+    if starting_grid_shape is None:
+        starting_grid_shape = [1, 1, 1]
+
+    # TODO: we can generalize to 2D points. 1D points do not make much sense
+    xyz = points[["x", "y", "z"]].values
+    tree = KDTree(xyz)
+
+    mins = np.min(xyz, axis=0)
+    maxs = np.max(xyz, axis=0)
+
+    remaining_indices = set(range(len(xyz)))
+
+    grid: dict[int, GridLevel] = {}
+    grid_level = GridLevel(
+        level=0,
+        grid_shape=starting_grid_shape,
+        mins=mins,
+        maxs=maxs,
+        limit=limit,
+        parent_cells=[(0, 0, 0)],
+        parent_grid_shape=starting_grid_shape,
+    )
+    # to avoid the risk of points in the boundary of the grid not being included
+    eps = 1e-6
+    previous_remaining_indices = len(remaining_indices)
+
+    while len(remaining_indices) > 0:
+        # initialization
+        grid[grid_level.level] = grid_level
+        if PRINT_DEBUG or True:
+            print(
+                f"Processing grid level {grid_level.level} with shape {grid_level.grid_shape} "
+                f"and chunk size {grid_level.chunk_size}. Remaining points: {len(remaining_indices)}"
+            )
+
+        # main logic
+        if PRINT_DEBUG:
+            print("Active cells: ", grid_level.cells)
+        for i, j, k in grid_level.iter_cells():
+            # calculate the centroid of the grid cell
+            centroid = grid_level.centroid((i, j, k))
+
+            # find points in the grid cell
+            # this filter points by a radius r, but we have different values per axis
+            indices = tree.query_ball_point(
+                centroid, r=grid_level.chunk_size.max().item() / 2 + eps, p=np.inf
+            )
+            filtered = xyz[indices]
+            mask = (
+                (centroid[0] - grid_level.chunk_size[0] - eps <= filtered[:, 0])
+                & (filtered[:, 0] <= centroid[0] + grid_level.chunk_size[0] + eps)
+                & (centroid[1] - grid_level.chunk_size[1] - eps <= filtered[:, 1])
+                & (filtered[:, 1] <= centroid[1] + grid_level.chunk_size[1] + eps)
+                & (centroid[2] - grid_level.chunk_size[2] - eps <= filtered[:, 2])
+                & (filtered[:, 2] <= centroid[2] + grid_level.chunk_size[2] + eps)
+            )
+            discarded = np.sum(~mask).item()
+            if discarded > 0:
+                # TODO: possible bug! This message is not printed while I would
+                #  expect that the kDTree query would return more points than the mask
+                #  would allow (this should happend when chunk_size has different
+                #  dimensions
+                if PRINT_DEBUG or True:
+                    print(
+                        f"-----------------> {discarded} points where filtered out of"
+                        f" {len(indices)}"
+                    )
+            indices = np.array(indices)[mask].tolist()
+
+            # filter out points that are not in the grid cell
+            indices = [i for i in indices if i in remaining_indices]
+
+            if len(indices) > 0:
+                if len(indices) <= limit:
+                    emitted = indices
+                else:
+                    emitted = RNG.choice(indices, size=limit, replace=False)
+                if PRINT_DEBUG:
+                    print(
+                        f"Emitting {len(emitted)} points for grid cell ({i}, {j}, {k})"
+                    )
+                grid_level.populated_cells[(i, j, k)] = xyz[emitted]
+                remaining_indices.difference_update(emitted)
+
+                # create a new layer for this grid cell
+                # layer_name = f"level_{grid_level.level}_cell_{i}_{j}_{k}"
+
+                # here we would save the points to the precomputed format
+                # e.g., save_points_to_precomputed(points[indices], precomputed_path, layer_name)
+
+        # np.take(xyz, indices, axis=0)
+        remaining_xyz = xyz[list(remaining_indices)]
+
+        # visual debug
+        VISUAL_DEBUG = False
+        if VISUAL_DEBUG:
+            plt.figure(figsize=(10, 10))
+            chunk_size = grid_level.chunk_size
+            lines_x = np.arange(
+                grid_level.mins[0],
+                grid_level.maxs[0] + chunk_size[0],
+                chunk_size[0] + eps,
+            )
+            lines_y = np.arange(
+                grid_level.mins[1],
+                grid_level.maxs[1] + chunk_size[1],
+                chunk_size[1] + eps,
+            )
+            for x in lines_x:
+                plt.plot(
+                    [x, x],
+                    [grid_level.mins[1], grid_level.maxs[1]],
+                    color="red",
+                    linewidth=0.5,
+                )
+            for y in lines_y:
+                plt.plot(
+                    [grid_level.mins[0], grid_level.maxs[0]],
+                    [y, y],
+                    color="red",
+                    linewidth=0.5,
+                )
+
+            plt.scatter(
+                remaining_xyz[:, 0],
+                remaining_xyz[:, 1],
+                s=100,
+                c=remaining_xyz[:, 2],
+            )
+            if len(remaining_xyz) > 0:
+                cbar = plt.colorbar()
+                cbar.set_ticks([remaining_xyz[:, 2].min(), remaining_xyz[:, 2].max()])
+                cbar.set_ticklabels(
+                    [f"{remaining_xyz[:, 2].min()}", f"{remaining_xyz[:, 2].max()}"]
+                )
+            plt.show()
+
+        # prepare for the next level
+        grid_level = GridLevel(
+            level=grid_level.level + 1,
+            grid_shape=grid_level.get_next_grid_shape(),
+            mins=grid_level.mins,
+            maxs=grid_level.maxs,
+            limit=limit,
+            parent_cells=list(grid_level.populated_cells.keys()),
+            parent_grid_shape=grid_level.grid_shape,
+        )
+
+        # sanity check
+        if len(remaining_indices) == previous_remaining_indices:
+            raise ValueError(
+                "No points were emitted in this grid level. This is likely due to the "
+                "grid size being too small."
+            )
+        previous_remaining_indices = len(remaining_indices)
+
+    print("spatial index computed, now saving to precomputed format")
+    return grid
